requirements-dev.txt
mock
ipython
django-extensions
pyzmq
jinja2
tornado
sphinx
